+++ index.html
@@ -228,7 +228,7 @@
     </div>
   </template>
 
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js" ></script>
 
 </body>
 </html>
