+++ js/constants.js
@@ -1,0 +1,41 @@
+const TYPES = ['palace', 'flat', 'house', 'bungalow'];
+const HOURS = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const IMAGES = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+const TITLES = [
+  'Pettyburg',
+  'КОТЭ',
+  'Милый друг',
+  'CAT INN',
+  'КОТВИЛЛЬ',
+  'ЗооОтель',
+  'Гранат',
+  'Лесси',
+  'BookingCat',
+  'Кошки в дом',
+];
+const DESCRIPTIONS = [
+  'Лучшая гостиница для животных в Санкт-Петербурге для кошек, собак, хорьков, кроликов, морских свинок, грызунов, птиц.',
+  'Гостиница для животных КОТЭ предлагает одни из самых низких цен в городе, удобный трансфер.',
+  'В нашем отеле для собак и кошек всегда открыты двери для питомцев, чьи хозяева какое-то время не могут осуществлять уход за любимцем. Помимо присмотра и ухода за четвероногим другом, мы оказываем весь спектр услуг, направленный на обеспечение комфорта и удобства для наших гостей и их хозяев.',
+  'Забудьте про волнение и переживания за ваших питомцев во время отъезда',
+  'Большие уютные номера с системой вентиляции. Одни из самых больших в Санкт-Петербурге.',
+  'Наша зоогостиница - это хорошее место, где Вы спокойно можете оставить своего питомца, если Вам необходимо уехать в отпуск, командировку, сделать ремонт в квартире, да мало ли, что еще...',
+  'Зоогостинница "Гранат" - это гостинница для собак и кошек премиум класса',
+  'Работаем с 1998 года. Клубная частная  структура организации качественного, полноценного, курортного отдыха для животных в экологически чистой зоне. Восполняем то, чего так не хватает в городе: свежий воздух, разнообразные прогулки, игры и плавание.',
+  'BookingCat – первая уникальная сеть отелей для питомцев. Мы предоставляем услуги по проживанию питомцев на период отпуска, командировки, болезни, ремонта и переезда владельца. Наши гости - это котики, собачки мелких пород, кролики, грызуны, птицы, и рыбки.',
+  'Уютная гостиница для домашних животных «Кошки в Дом» расположилась центре Петербурга. Мы создавали душевное место для размещения домашних любимцев, где питомец будет чувствовать себя комфортно, пока хозяева отдыхают на море или находятся в командировке.',
+];
+
+export {
+  TYPES,
+  HOURS,
+  FEATURES,
+  IMAGES,
+  TITLES,
+  DESCRIPTIONS
+};
